created DiffTypes constants file, initial stab at rendered diff iframes
