+++ src/components/select-diffType.tsx
@@ -1,16 +1,5 @@
 import * as React from 'react';
-
-export interface DiffTypes {
-  [propName: string]: string;
-}
-
-export const diffTypes:DiffTypes = {
-  HIGHLIGHTED_TEXT: 'highlighted text',
-  HIGHLIGHTED_SOURCE: 'highlighted source',
-  HIGLIGHTED_RENDERED: 'higlighted rendered', 
-  SIDE_BY_SIDE_RENDERED: 'side-by-side rendered', 
-  SIDE_BY_SIDE_TEXT: 'side-by-side text',
-}
+import {diffTypes} from '../constants/DiffTypes';
 
 export default class SelectDiffType extends React.Component<any, any> {
